[64] Add the ability to update the value of a non-containment reference when it is deserialized
Bug: https://github.com/eclipse-sirius/sirius-emf-json/issues/64
