+++ org.eclipse.sirius.emfjson/src/test/resources/unit/references/migration/LibraryWithContainmentAndNonContainmentEReference.xmi
@@ -1,0 +1,21 @@
+<?xml version="1.0" encoding="ASCII"?>
+<xmi:XMI xmi:version="2.0"
+         xmlns:xmi="http://www.omg.org/XMI"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xmlns:extlib="http:///org/eclipse/emf/examples/library/extlibrary.ecore/1.0.0"
+         xsi:schemaLocation="http:///org/eclipse/emf/examples/library/extlibrary.ecore/1.0.0 extlibrary.ecore">
+    <extlib:Library
+        name="library one">
+            <customBranches
+                name="library three">
+              <stock
+                  xsi:type="extlib:Book"
+                  title="book one"
+                  library="/0/@customBranches.0"
+                  libraries="/0 /1 /0/@customBranches.1"/>
+            </customBranches>
+            <customBranches name="library four"/>
+    </extlib:Library>
+    <extlib:Library
+            name="library two"/>
+</xmi:XMI>
